Update ggreg20_v3_drv.be
Updated the code and added an interval counter (x / 60 seconds) to the UI.

Also added the GGreg20_V3 power of the previous minute in µSv/h (var pwr_minute) to the UI and MQTT.
I didn't want to save the ma5 value to my database. This way I can grab and save the power value to my database every minute. Grafana will then calculate the moving average for me if needed.

There was also an error in calculating the dose.

The dose was calculated as follows:

dose = dose + (self.num_ctr_val * 0.00009)

Where 0.00009 is the factor of 0.0054 devided by 60 minutes.

However, self.num_ctr_val was already calculated with the factor of 0.0054:

self.num_ctr_val = cpm * 0.0054

Because of that, the calculated dose was waaaay too low (factor 0.0054 was used 2 times).

My updated code uses:
dose = dose + (pwr_minute / 60)